restriction, emails, cgu, cgv

+++ src/Controller/ProfileUserController.php
@@ -18,38 +18,55 @@ use Symfony\Component\Routing\Annotation\Route;
  */
 class ProfileUserController extends AbstractController
 {
-    /**
-     * @Route("/", name="app_profile_user_index", methods={"GET"})
-     */
-    public function index(UserRepository $userRepository): Response
-    {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');       
+    // /**
+    //  * @Route("/", name="app_profile_user_index", methods={"GET"})
+    //  */
+    // public function index(UserRepository $userRepository): Response
+    // {
+    //     $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');       
 
-        return $this->render('profile_user/index.html.twig', [
-            'users' => $userRepository->findAll(),
-        ]);
+    //     return $this->render('profile_user/index.html.twig', [
+    //         'users' => $userRepository->findAll(),
+    //     ]);
+    // }
+
+
+    // accéder à l'espace personnel
+    /**
+     * @Route("/", name="app_profile_user_index")
+     */
+    public function index(): Response
+    {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+        return $this->render('profile_user/profile.html.twig', []);
     }
 
     /**
-     * @Route("/{id}", name="app_profile_user_show", methods={"GET"})
+     * @Route("/", name="app_profile_user_show", methods={"GET"})
      */
-    public function show(User $user): Response
+    public function show(): Response
     {
-        return $this->render('profile_user/show.html.twig', [
-            'user' => $user,
-        ]);
+        return $this->render('profile_user/show.html.twig', []);
     }
 
     /**
-     * @Route("/{id}/edit", name="app_profile_user_edit", methods={"GET", "POST"})
+     * @Route("/modifier-mes-informations/{id}", name="app_profile_user_edit", methods={"GET", "POST"})
      */
     public function edit(Request $request, User $user, UserRepository $userRepository): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $form = $this->createForm(UserType::class, $user);
+        $form->remove('roles');
+        $form->remove('password');
+        $form->remove('isVerified');
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
             $userRepository->add($user);
+
             return $this->redirectToRoute('app_profile_user_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -60,14 +77,26 @@ class ProfileUserController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="app_profile_user_delete", methods={"POST"})
+     * @Route("/supprimer-mon-compte/{id}", name="app_profile_delete", methods={"GET"})
+     */
+    public function deleteProfile()
+    {
+        return $this->render('profile_user/confirm_delete.html.twig', []);
+    }
+
+    /**
+     * @Route("/confirmation-de-suppression/{id}", name="app_profile_user_delete", methods={"POST"})
      */
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+
             $userRepository->remove($user);
         }
 
-        return $this->redirectToRoute('app_profile_user_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_home');
+        // exit;
     }
 }
